+++ ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/api/rank/application/service/HashTagRankService.kt
@@ -1,20 +1,12 @@
 package backend.team.ahachul_backend.api.rank.application.service
 
+import backend.team.ahachul_backend.common.utils.FileUtils
 import backend.team.ahachul_backend.common.client.RedisClient
-import backend.team.ahachul_backend.common.exception.CommonException
 import backend.team.ahachul_backend.common.logging.Logger
-import backend.team.ahachul_backend.common.response.ResponseCode
 import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.google.gson.Gson
 import com.google.gson.JsonObject
 import org.springframework.stereotype.Service
-import java.io.BufferedWriter
-import java.io.FileOutputStream
-import java.io.IOException
-import java.io.OutputStreamWriter
-import java.nio.file.Files
-import java.nio.file.Path
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 
@@ -33,26 +25,15 @@ class HashTagRankService(
         val hashtagFileUrl = BASE_DIR + "${localDateTime.year}_${localDateTime.monthValue}_hashtag_log"
 
         // 저장 날짜에 해당하는 날짜 파일이 없다면 생성
-        createFile(hashtagFileUrl)
+        FileUtils.createFile(hashtagFileUrl)
 
         // 임시 버퍼에 들어오는 데이터 저장
         val jsonObject = createJsonObject(name, userId, localDateTime)
         if (buffer.size >= MAX_BUFFER_SIZE) {
-            saveFile(hashtagFileUrl, buffer)
+            FileUtils.saveFile(hashtagFileUrl, buffer)
             buffer = mutableListOf()
         }
         buffer.add(jsonObject)
-    }
-
-    private fun createFile(fileUrl: String) {
-        val path = Path.of(fileUrl).normalize()
-        try {
-            if (!Files.exists(path)) {
-                Files.createFile(path)
-            }
-        } catch (ex: IOException) {
-            throw CommonException(ResponseCode.FILE_CREATE_FAILED, ex)
-        }
     }
 
     private fun createJsonObject(name: String, userId: String, localDateTime: LocalDateTime): JsonObject {
@@ -62,23 +43,6 @@ class HashTagRankService(
         jsonObject.addProperty("userId", userId)
         jsonObject.addProperty("timestamp", currentTime)
         return jsonObject
-    }
-
-    private fun saveFile(filePath: String, data: List<Any>) {
-        val gson = Gson()
-        logger.info("save hashtag log in to file : $filePath")
-        try {
-            BufferedWriter(OutputStreamWriter(FileOutputStream(filePath, true), "UTF-8")).use {
-                data
-                    .map { data -> gson.toJson(data) }
-                    .forEach { str ->
-                        it.write(str)
-                        it.newLine()
-                    }
-            }
-        } catch (ex: IOException) {
-            throw CommonException(ResponseCode.FILE_WRITE_FAILED, ex)
-        }
     }
 
     /**
